requirements_datacorpus.txt
pandas==2.2.1
python-dotenv==1.0.1
pymongo==4.6.2
requests==2.31.0
beautifulsoup4==4.12.3
colorama==0.4.6
markdownify==0.11.6
tiktoken